% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{find_collision}
\alias{find_collision}
\title{Run `memorid()` until a collision is detected}
\usage{
find_collision()
}
\value{
the number of runs taken to reach the collision
}
\description{
Returns the number of runs taken to reach the collision
}
\keyword{internal}
