% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{mean_time_before_collision}
\alias{mean_time_before_collision}
\title{repeatedly runs `find_collision()` and returns the mean of the results}
\usage{
mean_time_before_collision(num = 100)
}
\arguments{
\item{num}{the number of `find_collision()` runs to perform. Default: 100}
}
\value{
the mean of multiple `find_collision()` results
}
\description{
Returns the mean of the `find_collision()` results
}
\keyword{internal}
