% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorid.R
\name{memorid}
\alias{memorid}
\title{Get a memorable ID}
\usage{
memorid(sep = "-")
}
\arguments{
\item{sep}{the seperator. Deafult is "-"}
}
\value{
a character string
}
\description{
Returns a memorable ID string
}
