% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorids.R
\docType{package}
\name{memorids}
\alias{memorids}
\alias{memorids-package}
\title{memorids: A package for creating memorable IDs.}
\description{
Sometimes you just need an easy to remember ID. memorids aims to give you
just that, by providing memorable strings that can be used for a variety of
purposes.
}
\details{
The simple strings are two English words, output in the format
"adjective-noun". The seperator is configurable.
}
\section{Uses}{
 A memorid can be used wherever you like, but remember that
  there are a small (just over half a million) number of combinations, so in
  mission critical settings the likelihood of the same string being returned
  twice is high. If it's important to you that this not happen, either use
  something else entirely, or take other steps to mitigate the risks.
}

\section{memorids functions}{
 The package only contains a single function,
  \code{memorid()} which returns memorable ID.
}

