% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_stats.R
\name{memorids_stats}
\alias{memorids_stats}
\title{print some basic info about the words used in this package}
\usage{
memorids_stats(num_runs = 100)
}
\arguments{
\item{num_runs}{the number of times to run the collision test}
}
\description{
print some basic info about the words used in this package
}
